Reorganize execution code
