+++ src/lib.rs
@@ -31,19 +31,17 @@ pub mod reference;
 pub mod ast;
 mod checker;
 mod execution;
-mod execution_error;
 pub mod functions;
 pub mod graph;
-mod lazy_execution;
 pub mod parse_error;
 mod parser;
 mod variables;
 
+pub use execution::error::ExecutionError;
 pub use execution::CancellationError;
 pub use execution::CancellationFlag;
 pub use execution::ExecutionConfig;
 pub use execution::NoCancellation;
-pub use execution_error::ExecutionError;
 pub use parser::Location;
 pub use parser::ParseError;
 pub use variables::Globals as Variables;
